requirements.txt
jupyter
jupyterlab
matplotlib
numpy
pandas
scikit-learn
scipy
tensorflow-gpu==2.0